changed double and some int to float

+++ include/minirt.h
@@ -46,8 +46,8 @@ typedef struct s_hitable
 	char			type;
 	t_vector3		pos;
 	t_vector3		normal;
-	double			diameter;
-	double			height;
+	float			diameter;
+	float			height;
 	unsigned int	color;
 	void			*next;
 }	t_hitable;
@@ -55,13 +55,13 @@ typedef struct s_hitable
 typedef struct s_light
 {
 	t_vector3	pos;
-	double		brightness;
+	float		brightness;
 	//unsigned int			color;
 }	t_light;
 
 typedef struct s_ambient
 {
-	double			brightness;
+	float			brightness;
 	unsigned int	color;
 }	t_ambient;
 
@@ -79,12 +79,12 @@ typedef struct s_data
 }				t_data;
 
 t_camera	create_camera(t_data *data, t_vector3 pos, t_vector3 normal, int FOV);
-t_ambient	create_ambient(double brightness, unsigned int color);
-t_light		create_light(t_vector3 pos, double brightness);
+t_ambient	create_ambient(float brightness, unsigned int color);
+t_light		create_light(t_vector3 pos, float brightness);
 void		add_hitable(t_hitable **list, t_hitable hit);
 t_hitable	create_sphere(t_vector3 pos, float diameter, unsigned int color);
 t_hitable	create_plane(t_vector3 pos, t_vector3 normal, unsigned int color);
-t_hitable	create_cylinder(t_vector3 pos, t_vector3 normal, int diameter, int height, unsigned int color);
+t_hitable	create_cylinder(t_vector3 pos, t_vector3 normal, float diameter, float height, unsigned int color);
 void		draw(t_data *data);
 void		keydown(void *data);
 void		mouse_hook(void *data);
