+++ src/light.c
@@ -6,13 +6,13 @@
 /*   By: bsyvasal <bsyvasal@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/06 01:04:27 by bsyvasal          #+#    #+#             */
-/*   Updated: 2024/04/06 01:42:55 by bsyvasal         ###   ########.fr       */
+/*   Updated: 2024/04/06 02:06:45 by bsyvasal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_light	create_light(t_vector3 pos, double brightness)
+t_light	create_light(t_vector3 pos, float brightness)
 {
 	t_light	light;
 
@@ -21,7 +21,7 @@ t_light	create_light(t_vector3 pos, double brightness)
 	return (light);
 }
 
-t_ambient	create_ambient(double brightness, unsigned int color)
+t_ambient	create_ambient(float brightness, unsigned int color)
 {
 	t_ambient	ambient;
